refactor(wiki): extract find/replace + bulk-replace into findreplace.go
